+++ src/containers/Gurb/pages/Contract/ContractSummary.jsx
@@ -4,7 +4,8 @@ import { useTranslation } from 'react-i18next'
 
 import TextRecomendation from '../../components/TextRecomendation'
 import SomStepper from '../../components/SomStepper'
-import { ReviewField, ReviewTable } from '../../components/ReviewField'
+import ReviewTable from '../../components/review/ReviewTable'
+import ReviewPricesTable from '../../components/review/ReviewPrices'
 
 import Typography from '@mui/material/Typography'
 import Box from '@mui/material/Box'
@@ -16,13 +17,11 @@ import LocationOnOutlinedIcon from '@mui/icons-material/LocationOnOutlined'
 import SettingsOutlinedIcon from '@mui/icons-material/SettingsOutlined'
 import LocalPhoneOutlinedIcon from '@mui/icons-material/LocalPhoneOutlined'
 import CreditCardOutlinedIcon from '@mui/icons-material/CreditCardOutlined'
-import LocalOfferOutlinedIcon from '@mui/icons-material/LocalOfferOutlined'
 
 import {
   iconRequirements,
   textBody1,
-  textHeader4,
-  textReviewLabel
+  textHeader4
 } from '../../gurbTheme'
 import { getPrices } from '../../../../services/api'
 import { THOUSANDS_CONVERSION_FACTOR } from '../../../../services/utils'
@@ -159,62 +158,7 @@ const ContractSummary = (props) => {
     }
   ]
 
-  const Prices = ({ concept, name }) => {
-    let keys = concept ? Object.keys(concept) : []
-    keys.sort()
-
-    const differentValues = new Set(
-      Object.keys(concept).map((key) => concept[key]?.value)
-    )
-    if (differentValues.size === 1) {
-      for (const key in concept) {
-        return (
-          <ReviewField
-            label={`${name}`}
-            value={`${concept[key]?.value} ${concept[key]?.unit}`}
-          />
-        )
-      }
-    }
-    const labels =
-      keys.length === 2
-        ? [t('PEAK'), t('VALLEY')]
-        : keys.length === 3
-          ? [t('PEAK'), t('FLAT'), t('VALLEY')]
-          : keys
-
-    return (
-      <Box
-        sx={{
-          display: 'flex',
-          flexDirection: 'column'
-        }}>
-        {concept ? (
-          <Grid
-            container
-            sx={{
-              display: 'flex',
-              flexDirection: 'column'
-            }}>
-            {keys.map((key, index) => (
-              <Grid item key={index}>
-                <ReviewField
-                  label={`${labels[index]}:`}
-                  value={`${concept[key]?.value} ${concept[key]?.unit}`}
-                />
-              </Grid>
-            ))}
-          </Grid>
-        ) : (
-          <>{t('UNAVAILABLE')}</>
-        )}
-      </Box>
-    )
-  }
-
-  console.log('VALUES', values, loading)
   useEffect(() => {
-    console.log('useEffect!!')
     setLoading(true)
     let powerFields = Object.values(
       Object.fromEntries(
@@ -251,45 +195,7 @@ const ContractSummary = (props) => {
       <TextRecomendation title={t('GURB_CONTRACT_SUMMARY_TITLE')} />
       <SomStepper step={activeStep} connectors={7 + 1} />
       <ReviewTable tableFields={reviewFields} />
-
-      <Box
-        sx={{
-          display: 'flex',
-          flexDirection: 'column',
-          marginY: '2rem',
-        }}>
-        <Box
-          sx={{
-            display: 'flex',
-            alignItems: 'flex-start',
-            gap: '2rem'
-          }}>
-          <LocalOfferOutlinedIcon sx={iconRequirements} />
-          <Typography sx={textHeader4}>
-            {t('GURB_REVIEW_PRICES_POWER_TITLE')}
-          </Typography>
-        </Box>
-          <Grid container sx={{marginLeft: '5rem'}}>
-            {reviewPrices.map((detail, index) => {
-              return (
-                <Grid
-                  key={index}
-                  item
-                  xs={12}
-                  sm={6}
-                  sx={{
-                    marginY: '1rem'
-                  }}>
-                  <Typography sx={textReviewLabel}>
-                    {t(detail.title)}
-                  </Typography>
-                  <Prices concept={prices?.[`${detail.field}`]} />
-                </Grid>
-              )
-            })}
-          </Grid>
-      </Box>
-
+      <ReviewPricesTable reviewPrices={reviewPrices} prices={prices} />
       <Typography
         sx={{ ...textBody1, marginBottom: '1rem', marginTop: '4rem' }}
         variant="body2"
